+++ src/providers/strategies/LedgerProviderStrategy/helpers/getLedgerProvider/getLedgerProvider.ts
@@ -4,7 +4,9 @@ import { getAccountProvider } from 'providers/helpers/accountProvider';
 import { ledgerAccountSelector } from 'store/selectors/accountSelectors';
 import { ledgerLoginSelector } from 'store/selectors/loginInfoSelectors';
 import { getState } from 'store/store';
-import { getLedgerConfiguration } from './getLedgerConfiguration';
+import { getLedgerConfiguration } from './helpers/getLedgerConfiguration';
+import { initHwProvider } from './helpers/initHwProvider';
+import { LedgerConfigType } from '../../types/ledgerProvider.types';
 
 export async function getLedgerProvider(props?: {
   shouldInitProvider?: boolean; // provider will be initialized if not logged in
@@ -17,43 +19,25 @@ export async function getLedgerProvider(props?: {
 
   const ledgerProvider = new HWProvider();
 
-  const initHWProvider = async () => {
-    const hasAddressIndex = ledgerLogin?.index != null;
-
-    if (provider instanceof HWProvider && provider.isInitialized()) {
-      if (hasAddressIndex) {
-        await provider.setAddressIndex(ledgerLogin.index);
-      }
-
-      return provider;
-    }
-
-    if (shouldInitProvider) {
-      const isInitialized = await ledgerProvider.init();
-      if (!isInitialized) {
-        throw new Error('Failed to initialize Ledger Provider');
-      }
-    }
-
-    if (hasAddressIndex) {
-      await ledgerProvider.setAddressIndex(ledgerLogin.index);
-    }
-
-    return ledgerProvider;
-  };
-
   try {
     if (!shouldInitProvider && ledgerAccount) {
+      const ledgerConfig: LedgerConfigType = {
+        version: ledgerAccount.version,
+        dataEnabled: ledgerAccount.hasContractDataEnabled
+      };
       return {
         ledgerProvider,
-        ledgerConfig: {
-          version: ledgerAccount.version,
-          dataEnabled: ledgerAccount.hasContractDataEnabled
-        }
+        ledgerConfig
       };
     }
 
-    const initializedLedgerProvider = await initHWProvider();
+    const initializedLedgerProvider = await initHwProvider({
+      provider,
+      ledgerProvider,
+      ledgerLogin,
+      shouldInitProvider
+    });
+
     const ledgerConfig = await getLedgerConfiguration(
       initializedLedgerProvider
     );