+++ src/providers/strategies/LedgerProviderStrategy/helpers/index.ts
@@ -1,5 +1,4 @@
 export * from './getAuthTokenText';
-export * from './getLedgerConfiguration';
 export * from './getLedgerErrorCodes';
 export * from './getLedgerProvider';
 export * from './getLedgerVersionOptions';